Enable multiple peptide display (checkbox in GraphChromatogram properties). Take out asynchronous rendering of GraphChromatogram -- too many problems. Add Zedgraph as a subproject of Skyline, Topograph, and SeeMS.
Former-commit-id: aca5bb95d9ac640e66881b1533c5de8ee154fe12

+++ pwiz_tools/Skyline/Util/UtilIO.cs
@@ -606,15 +606,21 @@ namespace pwiz.Skyline.Util
                 {
                     string backupFile = GetBackupFileName(pathDestination);
                     FileEx.SafeDelete(backupFile, true);
+
                     // First try replacing the destination file, if it exists
-                    File.Replace(pathTemp, pathDestination, backupFile, true);
-                    FileEx.SafeDelete(backupFile, true);
+                    if (File.Exists(pathDestination))
+                    {
+                        File.Replace(pathTemp, pathDestination, backupFile, true);
+                        FileEx.SafeDelete(backupFile, true);
+                        return;
+                    }
                 }
                 catch (FileNotFoundException)
                 {
-                    // Or just move, if it does not.
-                    Helpers.TryTwice(() => File.Move(pathTemp, pathDestination));
                 }
+
+                // Or just move, if it does not.
+                Helpers.TryTwice(() => File.Move(pathTemp, pathDestination));
             }
         }
 
@@ -726,7 +732,8 @@ namespace pwiz.Skyline.Util
             {
                 try
                 {
-                    Helpers.TryTwice(() => File.Delete(path));
+                    if (path != null && File.Exists(path))
+                        Helpers.TryTwice(() => File.Delete(path));
                 }
 // ReSharper disable EmptyGeneralCatchClause
                 catch (Exception)
